feat!: Injecting check utility into Assertion
